Continued adding isEnabled properties to functions before realising it wasn't going to work. Made notes on what to do instead

+++ commands/create.js
@@ -1,19 +1,27 @@
-const fs = require('fs').promises;
 const { createSoundManifest } = require('../slav_utils');
 const { ADMIN_ID } = require('../config.json');
 
-module.exports = {
+const create = {
   name: 'create',
   description:
     'creates the sound_manifest.regularSounds array from filenames in the ./sounds directory',
-  execute: message => {
-    if (message.author.id === ADMIN_ID) {
-      createSoundManifest();
+  isEnabled: true,
+  execute: function(message) {
+    if (this.isEnabled) {
+      if (message.author.id === ADMIN_ID) {
+        createSoundManifest();
+      } else {
+        console.log('----------');
+        console.log('User is not admin');
+        message.author.send('This command is for admins only, blyat');
+        message.delete();
+      }
     } else {
       console.log('----------');
-      console.log('User is not admin');
-      message.author.send('This command is for admins only, blyat');
-      message.delete();
+      console.log('Command is disabled');
+      message.author.send('This command is disabled!');
     }
   }
 };
+
+module.exports = create;
